♻️ move jsonStringify to @datadog/js-core/util
Move `jsonStringify`, `detachToJsonMethod`, and the `ObjectWithToJsonMethod`
type from `browser-core/src/tools/serialisation/jsonStringify` to
`js-core/src/util/jsonStringify`. The spec file moves alongside.

The `noop` dependency (previously imported from browser-core's
`functionUtils`) is inlined as a private function to avoid pulling
browser-specific utilities into js-core.

All consumers (browser-core internals, browser-rum-core) now import
directly from `@datadog/js-core/util`.
