+++ packages/browser-core/src/domain/console/consoleObservable.ts
@@ -1,10 +1,9 @@
 import { clocksNow } from '@datadog/js-core/time'
-import { ConsoleApiName, globalConsole } from '@datadog/js-core/util'
+import { ConsoleApiName, globalConsole, jsonStringify } from '@datadog/js-core/util'
 import { callMonitored } from '@datadog/js-core/monitor'
 import { isError, computeRawError } from '../error/error'
 import { Observable, mergeObservables } from '../../tools/observable'
 import { sanitize } from '../../tools/serialisation/sanitize'
-import { jsonStringify } from '../../tools/serialisation/jsonStringify'
 import type { RawError } from '../error/error.types'
 import { ErrorHandling, ErrorSource, NonErrorPrefix } from '../error/error.types'
 import { computeStackTrace } from '../../tools/stackTrace/computeStackTrace'
