+++ packages/browser-core/src/domain/error/error.ts
@@ -1,8 +1,7 @@
 import type { ClocksState } from '@datadog/js-core/time'
-import { isIndexableObject } from '@datadog/js-core/util'
+import { isIndexableObject, jsonStringify } from '@datadog/js-core/util'
 import type { Context } from '@datadog/js-core/util'
 import { sanitize } from '../../tools/serialisation/sanitize'
-import { jsonStringify } from '../../tools/serialisation/jsonStringify'
 import type { StackTrace } from '../../tools/stackTrace/computeStackTrace'
 import { computeStackTrace } from '../../tools/stackTrace/computeStackTrace'
 import { toStackTraceString } from '../../tools/stackTrace/handlingStack'
