+++ packages/browser-core/src/domain/telemetry/telemetry.ts
@@ -1,5 +1,5 @@
 import { clocksNow } from '@datadog/js-core/time'
-import { getDebugMode, combine, globalObject, isWorkerEnvironment } from '@datadog/js-core/util'
+import { getDebugMode, combine, globalObject, isWorkerEnvironment, jsonStringify } from '@datadog/js-core/util'
 import type { Hook } from '@datadog/js-core/assembly'
 import type { RecursivePartial, Context } from '@datadog/js-core/util'
 import { DISCARDED } from '@datadog/js-core/assembly'
@@ -20,7 +20,6 @@ import { buildTags } from '../tags'
 import { BufferedObservable, Observable } from '../../tools/observable'
 import { display } from '../../tools/display'
 import { sendToExtension } from '../../tools/sendToExtension'
-import { jsonStringify } from '../../tools/serialisation/jsonStringify'
 import { NonErrorPrefix } from '../error/error.types'
 import type { StackTrace } from '../../tools/stackTrace/computeStackTrace'
 import { computeStackTrace } from '../../tools/stackTrace/computeStackTrace'
