+++ packages/browser-core/src/index.ts
@@ -158,7 +158,6 @@ export * from './tools/utils/timezone'
 export * from './tools/utils/numberUtils'
 export * from './tools/utils/objectUtils'
 export * from './tools/utils/functionUtils'
-export * from './tools/serialisation/jsonStringify'
 export * from './tools/serialisation/stringify'
 export * from './tools/utils/stringUtils'
 export * from './tools/matchOption'
