+++ packages/browser-core/src/transport/batch.ts
@@ -1,10 +1,9 @@
 import type { EndpointBuilder } from '@datadog/js-core/transport'
-import { MORE_DETAILS, computeBytesCount, ONE_KIBI_BYTE, mockable } from '@datadog/js-core/util'
+import { MORE_DETAILS, computeBytesCount, ONE_KIBI_BYTE, mockable, jsonStringify } from '@datadog/js-core/util'
 import type { Context } from '@datadog/js-core/util'
 import { DOCS_TROUBLESHOOTING, display } from '../tools/display'
 import { objectValues } from '../tools/utils/polyfills'
 import { isPageExitReason, createPageMayExitObservable } from '../browser/pageMayExitObservable'
-import { jsonStringify } from '../tools/serialisation/jsonStringify'
 import { createIdentityEncoder } from '../tools/encoder'
 import type { Encoder, EncoderResult } from '../tools/encoder'
 import type { Observable } from '../tools/observable'
