+++ packages/browser-rum-core/src/transport/formDataTransport.ts
@@ -1,6 +1,7 @@
 import type { Encoder, EncoderResult, DeflateEncoderStreamId } from '@datadog/browser-core'
 import type { Context, Uint8ArrayBuffer } from '@datadog/js-core/util'
-import { addTelemetryDebug, createHttpRequest, jsonStringify, objectEntries, ErrorSource } from '@datadog/browser-core'
+import { addTelemetryDebug, createHttpRequest, objectEntries, ErrorSource } from '@datadog/browser-core'
+import { jsonStringify } from '@datadog/js-core/util'
 import { clocksNow } from '@datadog/js-core/time'
 import { createEndpointBuilder } from '@datadog/js-core/transport'
 import type { RumConfiguration } from '../domain/configuration'
