+++ packages/js-core/src/entries/util.ts
@@ -37,3 +37,5 @@ export { ONE_KIBI_BYTE, ONE_MEBI_BYTE, computeBytesCount, concatBuffers } from '
 export type { Uint8ArrayBuffer } from '../util/byteUtils'
 export type { Context, ContextValue, ContextArray } from '../util/context'
 export { mockable, mockableReplacements } from '../util/mockable'
+export { jsonStringify, detachToJsonMethod } from '../util/jsonStringify'
+export type { ObjectWithToJsonMethod } from '../util/jsonStringify'
