+++ packages/js-core/src/util/jsonStringify.spec.ts
